refactor(app): remove ambient overlay window + monitor
Cuts the ambient overlay surface across both backend and frontend:

  Frontend:
   - delete src/components/AmbientOverlay.tsx + its test
   - drop ambient hash-route + import from src/main.tsx
   - drop ambient devtools exposure block from src/App.tsx
   - drop AmbientModeToggle + ambientMode hooks/UI from SettingsPage
   - trim src/lib/ambient.ts to only the icon-overlay helpers Task 7
     will remove

  Backend:
   - delete app/src/ambient/ (extract.rs, monitor.rs, types.rs, mod.rs)
   - move icon-overlay-relevant types into a new app/src/icon_types.rs
     (Task 7 will delete this file when it removes the icon overlay)
   - drop ambient window creation, ambient global shortcut, run_ambient_monitor
     spawn, and the 5 ambient Tauri commands in app/src/lib.rs
   - drop position_ambient_bottom_right + the ambient command surface
     in app/src/search.rs (rewire icon helpers to crate::icon_types)
   - drop ambient_mode + ambient_dismissals fields in app/src/state.rs

  Followup taken in scope (Task 7 prep that became blocking once
  app/src/ambient/ was deleted):
   - drop the unreachable DragSnip/FocusChange/TextSelected/TextIcon/HideIcon
     match arms in app/src/router/{intent.rs,bundle.rs}
   - drop the matching dead variants in app/src/trigger/types.rs
   - annotate dead-code helpers in app/src/sensor/vision.rs that were only
     used by snip/ambient with #[allow(dead_code)] so the strict pre-commit
     clippy gate passes (Task 7 will trim these once IconOverlay is gone)
