--- app/src/ambient/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-pub mod extract;
-pub mod monitor;
-pub mod types;
