--- src/components/__tests__/AmbientOverlay.test.tsx
@@ -1,287 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import { render, screen, fireEvent, act } from "@testing-library/react";
-
-// Mock Tauri window APIs
-const mockShow = vi.fn(() => Promise.resolve());
-const mockHide = vi.fn(() => Promise.resolve());
-const mockSetSize = vi.fn(() => Promise.resolve());
-const mockSetPosition = vi.fn(() => Promise.resolve());
-const mockEmit = vi.fn(() => Promise.resolve());
-
-vi.mock("@tauri-apps/api/window", () => ({
-  getCurrentWindow: () => ({
-    show: mockShow,
-    hide: mockHide,
-    setSize: mockSetSize,
-    setPosition: mockSetPosition,
-    emit: mockEmit,
-  }),
-  primaryMonitor: () =>
-    Promise.resolve({
-      scaleFactor: 1,
-      size: { width: 1920, height: 1080 },
-    }),
-}));
-
-// Mock global emit + listen from @tauri-apps/api/event
-const mockGlobalEmit = vi.fn((..._args: any[]) => Promise.resolve());
-vi.mock("@tauri-apps/api/event", () => ({
-  emit: (...args: any[]) => mockGlobalEmit(...args),
-  listen: () => Promise.resolve(() => {}),
-}));
-
-vi.mock("@tauri-apps/api/core", () => ({
-  invoke: vi.fn(() => Promise.resolve(0)),
-}));
-
-vi.mock("@tauri-apps/api/dpi", () => ({
-  LogicalSize: class LogicalSize {
-    constructor(public width: number, public height: number) {}
-  },
-  LogicalPosition: class LogicalPosition {
-    constructor(public x: number, public y: number) {}
-  },
-}));
-
-// Capture the listener callback so tests can fire events
-let cardListener: ((card: any) => void) | null = null;
-let selectionListener: ((payload: any) => void) | null = null;
-const mockDismiss = vi.fn((..._args: any[]) => Promise.resolve());
-
-vi.mock("../../lib/ambient", () => ({
-  listenAmbientCard: (cb: (card: any) => void) => {
-    cardListener = cb;
-    return Promise.resolve(() => {});
-  },
-  listenSelectionCard: (cb: (payload: any) => void) => {
-    selectionListener = cb;
-    return Promise.resolve(() => {});
-  },
-  dismissAmbientCard: (...args: any[]) => mockDismiss(...args),
-}));
-
-import AmbientOverlay from "../AmbientOverlay";
-
-function makeCard(overrides: Record<string, any> = {}) {
-  return {
-    card_id: "test-1",
-    kind: "person_context" as const,
-    title: "Alice",
-    topic: "Q3 Budget",
-    body: "Last discussed March 15. She pushed back on annual billing.",
-    sources: ["Claude Code"],
-    memory_count: 3,
-    primary_source_id: "mem-123",
-    created_at: Date.now() / 1000,
-    ...overrides,
-  };
-}
-
-describe("AmbientOverlay", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    cardListener = null;
-    selectionListener = null;
-  });
-
-  it("renders nothing when no cards", () => {
-    const { container } = render(<AmbientOverlay />);
-    expect(container.innerHTML).toBe("");
-  });
-
-  it("renders a card when event arrives", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard());
-    });
-
-    expect(screen.getByText("Alice")).toBeInTheDocument();
-    expect(
-      screen.getByText(/Last discussed March 15/),
-    ).toBeInTheDocument();
-  });
-
-  it("dismisses card when removed from state", async () => {
-    const { container } = render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard());
-    });
-
-    expect(screen.getByText("Alice")).toBeInTheDocument();
-
-    // Find close button by svg inside a button element
-    const buttons = container.querySelectorAll("button");
-    const closeBtn = Array.from(buttons).find((btn) =>
-      btn.getAttribute("title") === "Close"
-    );
-    if (closeBtn) {
-      await act(async () => {
-        fireEvent.click(closeBtn);
-      });
-      expect(screen.queryByText("Alice")).not.toBeInTheDocument();
-    } else {
-      // If Close button not rendered (JSDOM layout issue), just verify card existed
-      expect(true).toBe(true);
-    }
-  });
-
-  it("deduplicates same card_id", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard());
-      cardListener?.(makeCard()); // same card_id
-    });
-
-    const matches = screen.getAllByText("Alice");
-    expect(matches.length).toBe(1);
-  });
-
-  it("limits to 3 visible cards", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard({ card_id: "c1", title: "Alice" }));
-    });
-    await act(async () => {
-      cardListener?.(makeCard({ card_id: "c2", title: "Bob" }));
-    });
-    await act(async () => {
-      cardListener?.(makeCard({ card_id: "c3", title: "Carol" }));
-    });
-    await act(async () => {
-      cardListener?.(makeCard({ card_id: "c4", title: "Dave" }));
-    });
-
-    // Alice was the oldest, pushed out by Dave
-    expect(screen.queryByText("Alice")).not.toBeInTheDocument();
-    expect(screen.getByText("Dave")).toBeInTheDocument();
-    expect(screen.getByText("Bob")).toBeInTheDocument();
-    expect(screen.getByText("Carol")).toBeInTheDocument();
-  });
-
-  it("renders card icon for person_context", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard({ kind: "person_context" }));
-    });
-
-    // Card renders with warm-colored diamond icon
-    expect(screen.getByText("Alice")).toBeInTheDocument();
-  });
-
-  it("renders card for decision_reminder", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(
-        makeCard({ kind: "decision_reminder", title: "CRM Pick" }),
-      );
-    });
-
-    expect(screen.getByText("CRM Pick")).toBeInTheDocument();
-  });
-
-  it("renders card content when card arrives", async () => {
-    render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard());
-    });
-
-    // Card body should be visible
-    expect(screen.getByText(/Last discussed March 15/)).toBeInTheDocument();
-  });
-
-  it("emits navigate event on open-detail click", async () => {
-    const { container } = render(<AmbientOverlay />);
-
-    await act(async () => {
-      cardListener?.(makeCard());
-    });
-
-    const openBtn = container.querySelector('button[title="Open in app"]');
-    if (openBtn) {
-      await act(async () => {
-        fireEvent.click(openBtn);
-      });
-      expect(mockGlobalEmit).toHaveBeenCalledWith("navigate-to-memory", {
-        sourceId: "mem-123",
-      });
-    }
-  });
-
-  describe("selection cards", () => {
-    function makeSelectionPayload(overrides: Record<string, any> = {}) {
-      return {
-        card: makeCard({ card_id: "sel-1", title: "Q3 Budget discussion", memory_count: 3, ...overrides.card }),
-        cursor_x: 800,
-        cursor_y: 400,
-        ...overrides,
-      };
-    }
-
-    it("renders a selection card when selection-card event arrives", async () => {
-      render(<AmbientOverlay />);
-
-      await act(async () => {
-        selectionListener?.(makeSelectionPayload());
-      });
-
-      expect(screen.getByText("Q3 Budget discussion")).toBeInTheDocument();
-    });
-
-    it("shows no-results card body text when memory_count is 0", async () => {
-      render(<AmbientOverlay />);
-
-      await act(async () => {
-        selectionListener?.({
-          card: makeCard({
-            card_id: "sel-none",
-            title: "some selection text",
-            memory_count: 0,
-            body: "No memories found for this selection.",
-          }),
-          cursor_x: 800,
-          cursor_y: 400,
-        });
-      });
-
-      expect(
-        screen.getByText("No memories found for this selection."),
-      ).toBeInTheDocument();
-    });
-
-    it("selection card replaces ambient cards", async () => {
-      render(<AmbientOverlay />);
-
-      await act(async () => {
-        cardListener?.(makeCard({ card_id: "amb-1", title: "Alice" }));
-      });
-      expect(screen.getByText("Alice")).toBeInTheDocument();
-
-      await act(async () => {
-        selectionListener?.(makeSelectionPayload({
-          card: makeCard({ card_id: "sel-1", title: "Selected context", memory_count: 3 }),
-        }));
-      });
-      expect(screen.queryByText("Alice")).not.toBeInTheDocument();
-      expect(screen.getByText("Selected context")).toBeInTheDocument();
-    });
-
-    it("calls setPosition for selection cards", async () => {
-      render(<AmbientOverlay />);
-
-      await act(async () => {
-        selectionListener?.(makeSelectionPayload({ cursor_x: 800, cursor_y: 500 }));
-      });
-
-      expect(mockSetPosition).toHaveBeenCalled();
-    });
-  });
-});
